requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
python-dotenv==1.0.0
pytest==7.4.3
flake8==6.1.0
black==23.12.1
mlflow==2.9.2
dagshub==0.2.21
psycopg2-binary==2.9.9  # PostgreSQL支持